# Improve performance of converting of converting URLs to strings