--- a/CHANGES/1198.misc.rst
+++ b/CHANGES/1198.misc.rst
@@ -1,0 +1,1 @@
+Improved performance of converting :class:`~yarl.URL` objects to strings -- by :user:`bdraco`.