<!-- Thank you for your contribution! -->

## What do these changes do?

Improve performance of converting of converting URLs to strings
There was a performance regression here as shown by the benchmarks in #1197

before
```
Convert URL to string: 0.097 sec
Convert URL with path to string: 0.097 sec
Convert URL with query to string: 0.145 sec
```
after
```
Convert URL to string: 0.045 sec
Convert URL with path to string: 0.045 sec
Convert URL with query to string: 0.091 sec
```

1.9.4
```
Convert URL to string: 0.064 sec
Convert URL with path to string: 0.042 sec
Convert URL with query to string: 0.113 sec
```